networkctl: split networkctl.c into small pieces
No functional change, just refactoring.
