+++ src/network/networkctl-description.h
@@ -1,0 +1,4 @@
+/* SPDX-License-Identifier: LGPL-2.1-or-later */
+#pragma once
+
+int dump_description(int argc, char *argv[]);
