+++ src/network/networkctl-journal.h
@@ -1,0 +1,4 @@
+/* SPDX-License-Identifier: LGPL-2.1-or-later */
+#pragma once
+
+int show_logs(int ifindex, const char *ifname);
