+++ src/network/networkctl-list.h
@@ -1,0 +1,4 @@
+/* SPDX-License-Identifier: LGPL-2.1-or-later */
+#pragma once
+
+int list_links(int argc, char *argv[], void *userdata);
