+++ src/network/networkctl-status-link.h
@@ -1,0 +1,4 @@
+/* SPDX-License-Identifier: LGPL-2.1-or-later */
+#pragma once
+
+int link_status(int argc, char *argv[], void *userdata);
